refactor: break down send view into multiple files/components
